models/requirements.txt
absl-py
mock
numpy
rouge-score
sacrebleu
sentencepiece
tensorflow-text==1.15.0rc0
tensor2tensor==1.15.0
tfds-nightly
tensorflow-gpu==1.15.2
tensorflow-datasets==3.0.0
tfds-nightly==3.1.0.dev202006230105